Revert "SparkleShare: Add patch to allow install on camera-less devices."
This reverts commit 998d42fd4e84afa09a2803dc25d61437d042a6d5. There
doesn't seem to be any point in patching this:
a) It's already been in the repo a long time
b) If we issue another version of a package, with the exact same
version name and code, how is anybody supposed to know which is which?
(I realise in this case it wouldn't make much difference, but the
general principle is important)
c) There is a technical problem in that if a different version of what
should be the same apk is pushed to the repo, installations will be
rejected by the client (until an update is done) because they will see
that the file they download is not what it was supposed to be.
d) The developer says he is fixing it in the next release anyway.