new app Wiki Loves Monument
The Wiki Loves Monuments is a context to take photo of cultural
monuments around the world. It is held every year.
http://www.wikilovesmonuments.org/

The WLMMobile application let you take pictures of monuments, fill in
metadata and upload them to the Wikimedia servers under a free license.
You can eventually win a prize awarded by a jury.