Fix Xabber (description, non-debug, target android-10, build otr4j)
dalb8's review indicated that the description should be closer to what
is available on Google Play.  I've included the list of supported
protocols and dropped that fact the Free Software is a good thing.

He also recommended killing the debuggable=true line, which was done
through a sed prebuild command.

The default target of android-9 was available on my machine, but not
the f-droid build server, so use the suggested alternative target of
android-10.

Otr4j is distributed as source in a separate repo but was included
in binary form in this one.  For now, clone and build otr4j but in
the future an f-droid srclib will be used.