WorldMap: Clarify description.
Avoid tautology, make separate paragraphs fit better. Don't go too
deep in technical details (like API levels; arguably, it's still
too technical, but I for example treat this app as an API usage example,
if we had multiple categories, it would have "Development" atatched).
Mention ability to serve as a standalone image viewer (which may come
as surprise to a user - suddenly, when clicking on a photo in a file
manager, popup appears with something called "WorldMap" in it).