new deactivated com.asksven.betterbatterystats
was able to make a very easy patch to remove the Locale-API integration
that was stopping the app from building. Just going to leave it for a
while, while we decide whether it's safe to publish. I haven't checked
the jars and while all the source code has Apache headers there doesn't
seem to be an overall licence.