deactivated 4.3.2 of com.google.zxing.client.android ...
changed to git-svn but I didn't verify (used svn myself) and set
UCM to RM. Not tags because I don't really understand the tagging
scheme—it is broader than just the android app. Note that the pom for
core takes care of copying the library after packaging. I've left it
deactivated so that we can discuss the anti-features and search plugs.