disable dashclock k9 apps ;replace with one new package ...
reasons are to do with fact that our k-9 build has a different package
name and any extensions we have published thus far only work with the
developer build. To help people using both packages we fork the package
name for the dashclock k9 app and remove the less fully featured one,
for simplicity.