Adaway: more WIP -- wait until gradle builds are solid ...
Jars are now built from source with ant (as recommended in the
Root Commands Readme); the alternative way would
have been to create build.gradles for them and edit build.gradle (and
possibly settings.gradle) to include them as dependencies but this isn't
too easy with our setup. Using local
aars (compiled android libraries) is apparently not recommended but
perhaps would be ok for a final build. The htmlcleaner jar doesn't
appear to be same as the canonical version so we pull that as an extlib
as in previous builds.