Mythdroid 0.6.3
Mythdroid 0.6.2 checks a server and determines that 0.6.3 is available.
Looking at the binaries provided at https://code.google.com/p/mythdroid/downloads/list
it appears that the release was made on May 12, 2013.  The SVN revision
valid as of that date is 918.  Also, with the new additions to features
the list at https://code.google.com/p/mythdroid/ was updated, and so
the updated list has been included as the f-droid Description.

android-13 was chosen since it was the most recent version that would
cleanly compile.