Issue #70: Fix Berlin vegan black screen issue
This commit actually does not fix the issue, it is just
a (not yet working) suggestion for fixing it:

The Berlin vegan guide needs a preparation phase, which inlucdes
copying some JS files and resources and stuff like that. This
preperation is done via

    ant prod.android

in the very root of the Berlin vegan guide git repository.

As you can see in the build.xml file in the root of the repo, the
target prod.android depends on the target prod, which in turn depends
on the target js.minify:

>     <target name="js.minify" depends="js.concatenate">
>
>         <apply executable="java" parallel="false">
>
>             <fileset dir="${build.dir}" includes="index.js"/>
>             <arg line="-jar"/>
>             <arg path="lib/tools/ycompressor.jar"/>
>             <srcfile/>
>             <arg line="-o"/>
>             <mapper type="glob" from="*.js"
>             to="${build.dir}/*-min.js"/>
>             <targetfile/>
>
>         </apply>
>         <delete file="${build.dir}/index.js"/>
>
>     </target>
>
> </project>

In the js.minify phase, the ycompressor.jar tool is used to compress
the java script files for mobile use.

By removing all the jar files in the prebuild phase, ycompressor.jar
gets removed, as well. Thus, the ant prod.android step does not work
properly.

I already managed to build this YUI compressor tool by hand from the sources.
But I could not figure out how to correctly add this external (non-Android)
host tool as dependency for this package.

Do you have any hint?

If I do it like this (in this commit) the build fails with:

> Running 'prebuild' commands in build/org.berlin_vegan.bvapp/platforms/android
> Cleaning subproject ../../../srclib/Cordova/framework
> Could not build app org.berlin_vegan.bvapp due to BuildException: Failed to update project at ../../../srclib/YUICompressor
> ==== detail begin ====
> Error: ../../../srclib/YUICompressor is not a valid project (AndroidManifest.xml not found).