Add version 2.0.3 of Berlin Vegan Guide
First of all we use the correct version name, version code, and commit
id.

We do not need the sed-command anymore: I patched the fdroidserver (see
commit dd0b38a), that if the version name is not found for some reason,
it just uses the git tag. That is exactly, what we want.

Removing the sed-command has one other advantage: the git repository has
no changes anymore or in other words it is not 'dirty' anymore. So there
is no '-dirty' string attached to the version string during our gradle
build process.

Also add Archive Policy. Only the newest version is going to build,
the other builds are referring to the old github repository and fail.