Fix UnitConverterUltimate build
The build was failing because some of the resources needed to support
donations (license key and whatnot) aren't included in the source.

This commit fixes the build by deleting the classes that required those
resources, as well as adding a prebuild command that removes the donate
option from the app's preference menu and manifest.