New app: F-Droid Privileged Extension OTA update ZIP
This is not actually an APK, but an OTA (Over-The-Air) update ZIP file that
is meant to be flashed onto the /system partition using the recovery. This
is the first non-APK to be built and included in the f-droid.org repo.

This relies on the buildserver running fdroidserver that includes:
https://gitlab.com/fdroid/fdroidserver/merge_requests/181

This should use the exact same signing key as org.fdroid.fdroid.privileged,
but the APK signing key can change for the update.zip, since upgrades do
not happen via the normal method, and instead the update.zip will remove
the current APK and replace it with the current one without a signature
check.
https://gitlab.com/fdroid/fdroidserver/issues/233

closes https://gitlab.com/fdroid/privileged-extension/issues/9