gitlab-ci: show diff when there are lint failures
Oftentimes, the lint failures are just simple whitespace issues.  Many
people submitting metadata do not really know much about the fdroid tools,
so they have to look up or ask what rewritemeta is.  This just shows a diff
after running rewritemeta so it should be clearer what the issue is.