gitlab-ci: if running on a fork, only check the files in the fork
To avoid confusing newbies, the CI should only run on the files that were
changed in the fork.  Sometimes lint/rewrite errors slip into the git repo,
and the CI for all merge requests will fail until those are fixed.  This
changes that to make the merge request CI only fail if the files in the
merge request are the ones with problems.

The CI runs on the main repo should will check everything.