gitlab-ci: fix creation of 'fdroid build' job
This should be run on all branches in forks and all merge requests, as
long as they change a metadata file.  Unfortunately gitlab-ci's "changes"
logic is not as sophisticated as the logic written in the job for detecting
which files changed.  gitlab-ci's logic only detects changes that happened
in the current push event, this job's logic compares to master.

Using gitlab-ci's "changes" logic means that if someone changed a metadata
file in a merge request, "fdroid build" would be triggered.  If they then
updated the merge request with a change to another file, then the next run
would not include a "fdroid build" job, since no metadata file was changed
in that push.