Update Local-GSM-Backend app
This apk doesn't work since over a month, because mozilla changed their
URL that is used to download their database and the application is
useless without database, see
https://github.com/n76/Local-GSM-Backend/issues/132

And from what you can read eg. in his readme file, the author is not
interested in working on this app anymore and doesn't even use it.

I recently updated the url myself, added a possibility to configure that
url in settings and also made the app working on my dual sim phone.

I think that I can maintain this app, because I'm going to use it in
near future, so I need it working anyway.