Add srclibs for RevoFonto and VokoGrundo
This repos contain the data for the Reta Vortaro dictonary used by
PReVo. Previously the data was stored in the Revo repo which was updated
manually by downloading tarballs from the Reta Vortaro website and
creating commits with the contents. In the meantime the official data
source for Reta Vortaro has migrated from an internal CVS repo to a
public github repo so PReVo can now use that directly instead of the
tarball repo.