gitlab-ci: use barebones base image for easy auditing
F-Droid setups here deliberately use simple base images like
alpine so we can easily check whether they are all free
software, like the MapBox issues.  I have done quite a bit of
profiling in the past of docker images like curlimages vs
installing.  For short jobs like this, there has never been a
clear winner.  When curlimages is in the runner's docker
cache, it can be a bit faster than installing the package, but
images like alpine are much more likely to be in the docker
cache, making them on average faster.

this reverts a change in !11503.