LibreOffice Viewer: Update to 7.6.5
Update to new release 7.6.5 and adjust the versionName
to just use the version number and drop the comment in the
maintainer notes about a part of the commit hash and the vendor
being part of the versionName.
This is no longer the case since LibreOffice commit [1],
which also removes one aspect previously making it impossible
to use the automatic update mechanism.
(The versionCode is still only set via an environment variable
during the build and not retrievable another way, so that still
prevents switching to the automatic update mechanism at the moment.)

[1] https://git.libreoffice.org/core/commit/35b7aa3a865eda90bec945ac2e11b20a75a37bd6